samples/reasoning_engine/requirements.txt
google-cloud-aiplatform[reasoningengine,langchain]==1.57.0
langchain-google-cloud-sql-pg==0.6.0
langchain-google-vertexai==1.0.1
google-cloud-resource-manager==1.12.3